moved autofiller to own module and into controller
